requirements.txt
discord.py
discord.py[voice]
pandas
pyttsx3